media/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.20
yt-dlp>=2024.12.13
requests>=2.32.5
numpy>=1.24.0
Pillow>=10.0.0
supabase>=2.11.0
vosk>=0.3.44
python-dotenv>=1.0.1
python-magic>=0.4.27
